requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==1.10.0
scikit-learn==1.3.2
numpy==1.26.4
python-telegram-bot==20.4
ccxt==4.5.7
Pillow==10.0.1
pytesseract
redis>=5.0.0
rq==1.16.2
pytest==7.4.0
alembic==1.11.1
flask-cors==4.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
